Refactor tests, remove fixtures
