+++ tabular/tests/unittests/calibrate/test_calibrate.py
@@ -1,4 +1,7 @@
-def test_calibrate_binary(fit_helper):
+from autogluon.tabular.testing import FitHelper
+
+
+def test_calibrate_binary():
     """Tests that calibrate=True doesn't crash in binary"""
     fit_args = dict(
         hyperparameters={"GBM": {}},
@@ -6,10 +9,10 @@ def test_calibrate_binary(fit_helper):
     )
     dataset_name = "toy_binary"
 
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
+    FitHelper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
 
 
-def test_calibrate_binary_bag(fit_helper):
+def test_calibrate_binary_bag():
     """Tests that calibrate=True doesn't crash in binary w/ bagging"""
     fit_args = dict(
         hyperparameters={"GBM": {"ag_args_ensemble": {"fold_fitting_strategy": "sequential_local"}}},
@@ -18,10 +21,10 @@ def test_calibrate_binary_bag(fit_helper):
     )
     dataset_name = "toy_binary"
 
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
+    FitHelper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
 
 
-def test_calibrate_multiclass(fit_helper):
+def test_calibrate_multiclass():
     """Tests that calibrate=True doesn't crash in multiclass"""
     fit_args = dict(
         hyperparameters={"GBM": {}},
@@ -29,10 +32,10 @@ def test_calibrate_multiclass(fit_helper):
     )
     dataset_name = "toy_multiclass"
 
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
+    FitHelper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
 
 
-def test_calibrate_multiclass_bag(fit_helper):
+def test_calibrate_multiclass_bag():
     """Tests that calibrate=True doesn't crash in multiclass w/ bagging"""
     fit_args = dict(
         hyperparameters={"GBM": {"ag_args_ensemble": {"fold_fitting_strategy": "sequential_local"}}},
@@ -41,10 +44,10 @@ def test_calibrate_multiclass_bag(fit_helper):
     )
     dataset_name = "toy_multiclass"
 
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
+    FitHelper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
 
 
-def test_calibrate_quantile(fit_helper):
+def test_calibrate_quantile():
     """Tests that calibrate=True doesn't crash in quantile"""
     fit_args = dict(
         hyperparameters={"RF": {}},
@@ -52,10 +55,10 @@ def test_calibrate_quantile(fit_helper):
     )
     dataset_name = "toy_quantile"
 
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
+    FitHelper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
 
 
-def test_calibrate_quantile_bag(fit_helper):
+def test_calibrate_quantile_bag():
     """Tests that calibrate=True doesn't crash in quantile w/ bagging"""
     fit_args = dict(
         hyperparameters={"RF": {"ag_args_ensemble": {"fold_fitting_strategy": "sequential_local"}}},
@@ -64,4 +67,4 @@ def test_calibrate_quantile_bag(fit_helper):
     )
     dataset_name = "toy_quantile"
 
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
+    FitHelper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
