+++ tabular/tests/unittests/dynamic_stacking/test_dynamic_stacking.py
@@ -4,6 +4,8 @@ import pytest
 
 from autogluon.core.constants import BINARY
 from autogluon.core.metrics import METRICS
+from autogluon.tabular.testing import FitHelper
+from autogluon.tabular.testing.fit_helper import stacked_overfitting_assert
 
 DS_ARGS_TEST_DEFAULTS = dict(
     validation_procedure="holdout",
@@ -17,7 +19,7 @@ DS_ARGS_TEST_DEFAULTS = dict(
 )
 
 
-def test_spot_and_avoid_stacked_overfitting(fit_helper, dataset_loader_helper):
+def test_spot_and_avoid_stacked_overfitting():
     """Tests that dynamic stacking works."""
     fit_args = dict(
         hyperparameters={"RF": {}, "GBM": {}},
@@ -33,7 +35,7 @@ def test_spot_and_avoid_stacked_overfitting(fit_helper, dataset_loader_helper):
     dataset_name = "adult"
     extra_metrics = list(METRICS[BINARY])
 
-    fit_helper.fit_and_validate_dataset(
+    FitHelper.fit_and_validate_dataset(
         dataset_name=dataset_name,
         fit_args=fit_args,
         extra_metrics=extra_metrics,
@@ -45,7 +47,7 @@ def test_spot_and_avoid_stacked_overfitting(fit_helper, dataset_loader_helper):
     )
 
 
-def test_dynamic_stacking_hps(fit_helper, dataset_loader_helper, stacked_overfitting_assert_func):
+def test_dynamic_stacking_hps():
     """Tests dynamic stacking arguments."""
     fit_args = dict(
         hyperparameters={"DUMMY": {}},
@@ -59,7 +61,7 @@ def test_dynamic_stacking_hps(fit_helper, dataset_loader_helper, stacked_overfit
     )
 
     # Get custom val data (the test data)
-    train_data, test_data, dataset_info = dataset_loader_helper.load_dataset(name="adult", directory_prefix="./datasets/")
+    train_data, test_data, dataset_info = FitHelper.load_dataset(name="adult", directory_prefix="./datasets/")
     label = dataset_info["label"]
     allowed_cols = ["age", label]
     train_data = train_data[allowed_cols]
@@ -81,19 +83,19 @@ def test_dynamic_stacking_hps(fit_helper, dataset_loader_helper, stacked_overfit
         tmp_fit_args = fit_args.copy()
         tmp_fit_args["ds_args"] = tmp_ds_args
         if expect_raise is None:
-            predictor = fit_helper.fit_dataset(train_data=train_data, init_args=dict(label=label), fit_args=tmp_fit_args, sample_size=1000)
+            predictor = FitHelper.fit_dataset(train_data=train_data, init_args=dict(label=label), fit_args=tmp_fit_args, sample_size=1000)
             if ("holdout_data" in ds_args_update) and (ds_args_update["holdout_data"] is not None):
                 n_expected = 1000 + n_test_data
                 assert len(predictor.predict_oof()) == n_expected, "Verify that holdout data was used for training"
             lb = predictor.leaderboard(test_data, extra_info=True)
-            stacked_overfitting_assert_func(lb, predictor, False, False)
+            stacked_overfitting_assert(lb, predictor, False, False)
             shutil.rmtree(predictor.path)
         else:
             with pytest.raises(expect_raise):
-                fit_helper.fit_dataset(train_data=train_data, init_args=dict(label=label), fit_args=tmp_fit_args, sample_size=1000)
+                FitHelper.fit_dataset(train_data=train_data, init_args=dict(label=label), fit_args=tmp_fit_args, sample_size=1000)
 
 
-def test_no_dynamic_stacking(fit_helper):
+def test_no_dynamic_stacking():
     """Tests that dynamic stacking does not run if stacking is disabled."""
     fit_args = dict(
         hyperparameters={"DUMMY": {}},
@@ -105,13 +107,13 @@ def test_no_dynamic_stacking(fit_helper):
     dataset_name = "adult"
     extra_metrics = list(METRICS[BINARY])
 
-    predictor = fit_helper.fit_and_validate_dataset(
+    predictor = FitHelper.fit_and_validate_dataset(
         dataset_name=dataset_name, fit_args=fit_args, extra_metrics=extra_metrics, expected_model_count=1, refit_full=False
     )
     assert predictor._stacked_overfitting_occurred is None
 
 
-def test_dynamic_stacking_fit_extra(fit_helper):
+def test_dynamic_stacking_fit_extra():
     """Tests that fit_extra works after dynamic stacking."""
     fit_args = dict(
         hyperparameters={"RF": {}},
@@ -126,7 +128,7 @@ def test_dynamic_stacking_fit_extra(fit_helper):
     dataset_name = "adult"
     extra_metrics = list(METRICS[BINARY])
 
-    predictor = fit_helper.fit_and_validate_dataset(
+    predictor = FitHelper.fit_and_validate_dataset(
         dataset_name=dataset_name,
         fit_args=fit_args,
         extra_metrics=extra_metrics,
@@ -150,7 +152,7 @@ def test_dynamic_stacking_fit_extra(fit_helper):
     shutil.rmtree(predictor.path, ignore_errors=True)
 
 
-def test_dynamic_stacking_with_time_limit(fit_helper):
+def test_dynamic_stacking_with_time_limit():
     """Tests that dynamic stacking does not run if stacking is disabled."""
     ds_args = DS_ARGS_TEST_DEFAULTS.copy()
     ds_args["holdout_frac"] = 0.5
@@ -168,7 +170,7 @@ def test_dynamic_stacking_with_time_limit(fit_helper):
     dataset_name = "adult"
     extra_metrics = list(METRICS[BINARY])
 
-    fit_helper.fit_and_validate_dataset(
+    FitHelper.fit_and_validate_dataset(
         dataset_name=dataset_name,
         fit_args=fit_args,
         extra_metrics=extra_metrics,
@@ -182,7 +184,7 @@ def test_dynamic_stacking_with_time_limit(fit_helper):
 
 
 @pytest.mark.timeout(120)  # if running AutoGluon twice fails due to a multiprocessing bug, we want to hang up and crash.
-def test_dynamic_stacking_run_twice_parallel_fold_fitting_strategy(fit_helper, dataset_loader_helper, stacked_overfitting_assert_func):
+def test_dynamic_stacking_run_twice_parallel_fold_fitting_strategy():
     """Tests that dynamic stacking memory save fit works."""
     ds_args = DS_ARGS_TEST_DEFAULTS.copy()
     ds_args["memory_safe_fits"] = True  # guarantee for sanity
@@ -198,14 +200,14 @@ def test_dynamic_stacking_run_twice_parallel_fold_fitting_strategy(fit_helper, d
     )
 
     # Get custom val data (the test data)
-    train_data, test_data, dataset_info = dataset_loader_helper.load_dataset(name="adult", directory_prefix="./datasets/")
+    train_data, test_data, dataset_info = FitHelper.load_dataset(name="adult", directory_prefix="./datasets/")
     label = dataset_info["label"]
     allowed_cols = ["age", label]
     train_data = train_data[allowed_cols]
     test_data = test_data[allowed_cols]
 
     for _ in range(2):
-        predictor = fit_helper.fit_dataset(train_data=train_data, init_args=dict(label=label), fit_args=fit_args, sample_size=1000)
+        predictor = FitHelper.fit_dataset(train_data=train_data, init_args=dict(label=label), fit_args=fit_args, sample_size=1000)
         lb = predictor.leaderboard(test_data, extra_info=True)
-        stacked_overfitting_assert_func(lb, predictor, False, False)
+        stacked_overfitting_assert(lb, predictor, False, False)
         shutil.rmtree(predictor.path)
