+++ tabular/tests/unittests/edgecases/test_edgecases.py
@@ -6,9 +6,10 @@ from pathlib import Path
 
 from autogluon.core.constants import BINARY
 from autogluon.core.metrics import METRICS
+from autogluon.tabular.testing import FitHelper
 
 
-def test_no_weighted_ensemble(fit_helper):
+def test_no_weighted_ensemble():
     """Tests that fit_weighted_ensemble=False works"""
     fit_args = dict(
         hyperparameters={"DUMMY": {}},
@@ -17,10 +18,10 @@ def test_no_weighted_ensemble(fit_helper):
     dataset_name = "toy_binary"
     extra_metrics = list(METRICS[BINARY])
 
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args, extra_metrics=extra_metrics, expected_model_count=1)
+    FitHelper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args, extra_metrics=extra_metrics, expected_model_count=1)
 
 
-def test_no_full_last_level_weighted_ensemble(fit_helper):
+def test_no_full_last_level_weighted_ensemble():
     """Tests that fit_weighted_ensemble=False works"""
     fit_args = dict(
         hyperparameters={"DUMMY": {}},
@@ -34,10 +35,10 @@ def test_no_full_last_level_weighted_ensemble(fit_helper):
     dataset_name = "toy_binary"
     extra_metrics = list(METRICS[BINARY])
 
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args, extra_metrics=extra_metrics, expected_model_count=4)
+    FitHelper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args, extra_metrics=extra_metrics, expected_model_count=4)
 
 
-def test_no_full_last_level_weighted_ensemble_additionally(fit_helper):
+def test_no_full_last_level_weighted_ensemble_additionally():
     """Tests that fit_weighted_ensemble=False works"""
     fit_args = dict(
         hyperparameters={"DUMMY": {}},
@@ -52,10 +53,10 @@ def test_no_full_last_level_weighted_ensemble_additionally(fit_helper):
     dataset_name = "toy_binary"
     extra_metrics = list(METRICS[BINARY])
 
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args, extra_metrics=extra_metrics, expected_model_count=4)
+    FitHelper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args, extra_metrics=extra_metrics, expected_model_count=4)
 
 
-def test_full_last_level_weighted_ensemble_additionally(fit_helper):
+def test_full_last_level_weighted_ensemble_additionally():
     """Tests that fit_weighted_ensemble=False works"""
     fit_args = dict(
         hyperparameters={"DUMMY": {}},
@@ -70,13 +71,13 @@ def test_full_last_level_weighted_ensemble_additionally(fit_helper):
     dataset_name = "toy_binary"
     extra_metrics = list(METRICS[BINARY])
 
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args, extra_metrics=extra_metrics, expected_model_count=5)
+    FitHelper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args, extra_metrics=extra_metrics, expected_model_count=5)
 
     fit_args["num_stack_levels"] = 0
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args, extra_metrics=extra_metrics, expected_model_count=2)
+    FitHelper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args, extra_metrics=extra_metrics, expected_model_count=2)
 
 
-def test_full_last_level_weighted_ensemble(fit_helper):
+def test_full_last_level_weighted_ensemble():
     """Tests that fit_weighted_ensemble=False works"""
     fit_args = dict(
         hyperparameters={"DUMMY": {}},
@@ -90,10 +91,10 @@ def test_full_last_level_weighted_ensemble(fit_helper):
     dataset_name = "toy_binary"
     extra_metrics = list(METRICS[BINARY])
 
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args, extra_metrics=extra_metrics, expected_model_count=4)
+    FitHelper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args, extra_metrics=extra_metrics, expected_model_count=4)
 
 
-def test_max_sets(fit_helper):
+def test_max_sets():
     """Tests that max_sets works"""
     fit_args = dict(
         hyperparameters={"DUMMY": {"ag_args_ensemble": {"max_sets": 3}}},
@@ -103,7 +104,7 @@ def test_max_sets(fit_helper):
     )
     dataset_name = "toy_binary"
 
-    predictor = fit_helper.fit_and_validate_dataset(
+    predictor = FitHelper.fit_and_validate_dataset(
         dataset_name=dataset_name,
         fit_args=fit_args,
         expected_model_count=1,
@@ -116,7 +117,7 @@ def test_max_sets(fit_helper):
     shutil.rmtree(predictor.path, ignore_errors=True)
 
 
-def test_num_folds(fit_helper):
+def test_num_folds():
     """Tests that num_folds works"""
     fit_args = dict(
         hyperparameters={"DUMMY": {"ag_args_ensemble": {"num_folds": 3}}},
@@ -126,7 +127,7 @@ def test_num_folds(fit_helper):
     )
     dataset_name = "toy_binary"
 
-    predictor = fit_helper.fit_and_validate_dataset(
+    predictor = FitHelper.fit_and_validate_dataset(
         dataset_name=dataset_name,
         fit_args=fit_args,
         expected_model_count=1,
@@ -139,7 +140,7 @@ def test_num_folds(fit_helper):
     shutil.rmtree(predictor.path, ignore_errors=True)
 
 
-def test_num_folds_hpo(fit_helper):
+def test_num_folds_hpo():
     """Tests that num_folds works"""
     fit_args = dict(
         hyperparameters={"GBM": {"ag_args_ensemble": {"num_folds": 2}}},
@@ -154,7 +155,7 @@ def test_num_folds_hpo(fit_helper):
     )
     dataset_name = "toy_binary"
 
-    predictor = fit_helper.fit_and_validate_dataset(
+    predictor = FitHelper.fit_and_validate_dataset(
         dataset_name=dataset_name,
         fit_args=fit_args,
         expected_model_count=2,
@@ -168,7 +169,7 @@ def test_num_folds_hpo(fit_helper):
     shutil.rmtree(predictor.path, ignore_errors=True)
 
 
-def test_use_bag_holdout_calibrate(fit_helper):
+def test_use_bag_holdout_calibrate():
     """
     Test that use_bag_holdout=True works for calibration
     Ensures the bug is fixed in https://github.com/autogluon/autogluon/issues/2674
@@ -183,7 +184,7 @@ def test_use_bag_holdout_calibrate(fit_helper):
     )
 
     dataset_name = "adult"
-    fit_helper.fit_and_validate_dataset(
+    FitHelper.fit_and_validate_dataset(
         dataset_name=dataset_name,
         init_args=init_args,
         fit_args=fit_args,
@@ -192,12 +193,12 @@ def test_use_bag_holdout_calibrate(fit_helper):
     )
 
 
-def test_num_folds_parallel(fit_helper, capsys):
+def test_num_folds_parallel(capsys):
     """Tests that num_folds_parallel equal to 1 works"""
     fit_args = dict(hyperparameters={"DUMMY": {}}, fit_weighted_ensemble=False, num_bag_folds=2, num_bag_sets=1, ag_args_ensemble=dict(num_folds_parallel=1))
     dataset_name = "toy_binary"
 
-    predictor = fit_helper.fit_and_validate_dataset(
+    predictor = FitHelper.fit_and_validate_dataset(
         dataset_name=dataset_name,
         fit_args=fit_args,
         expected_model_count=1,
@@ -209,12 +210,12 @@ def test_num_folds_parallel(fit_helper, capsys):
     shutil.rmtree(predictor.path, ignore_errors=True)
 
 
-def test_raises_num_cpus_float(fit_helper):
+def test_raises_num_cpus_float():
     """Tests that num_cpus specified as a float raises a TypeError"""
     fit_args = dict(num_cpus=1.0)
     dataset_name = "toy_binary"
     with pytest.raises(TypeError, match=r"`num_cpus` must be an int or 'auto'. Found: .*"):
-        fit_helper.fit_and_validate_dataset(
+        FitHelper.fit_and_validate_dataset(
             dataset_name=dataset_name,
             fit_args=fit_args,
             expected_model_count=None,
@@ -222,12 +223,12 @@ def test_raises_num_cpus_float(fit_helper):
         )
 
 
-def test_raises_num_cpus_zero(fit_helper):
+def test_raises_num_cpus_zero():
     """Tests that num_cpus=0 raises a ValueError"""
     fit_args = dict(num_cpus=0)
     dataset_name = "toy_binary"
     with pytest.raises(ValueError, match=r"`num_cpus` must be greater than or equal to 1. .*"):
-        fit_helper.fit_and_validate_dataset(
+        FitHelper.fit_and_validate_dataset(
             dataset_name=dataset_name,
             fit_args=fit_args,
             expected_model_count=None,
@@ -235,12 +236,12 @@ def test_raises_num_cpus_zero(fit_helper):
         )
 
 
-def test_raises_num_gpus_neg(fit_helper):
+def test_raises_num_gpus_neg():
     """Tests that num_gpus<0 raises a ValueError"""
     fit_args = dict(num_gpus=-1)
     dataset_name = "toy_binary"
     with pytest.raises(ValueError, match=r"`num_gpus` must be greater than or equal to 0. .*"):
-        fit_helper.fit_and_validate_dataset(
+        FitHelper.fit_and_validate_dataset(
             dataset_name=dataset_name,
             fit_args=fit_args,
             expected_model_count=None,
@@ -248,7 +249,7 @@ def test_raises_num_gpus_neg(fit_helper):
         )
 
 @pytest.mark.parametrize("delay_bag_sets", [True, False])
-def test_delay_bag_sets(fit_helper,delay_bag_sets):
+def test_delay_bag_sets(delay_bag_sets):
     """Tests that max_sets works"""
     fit_args = dict(
         hyperparameters={"DUMMY": [{}, {}]},
@@ -261,7 +262,7 @@ def test_delay_bag_sets(fit_helper,delay_bag_sets):
     )
     dataset_name = "toy_binary"
 
-    predictor = fit_helper.fit_and_validate_dataset(
+    predictor = FitHelper.fit_and_validate_dataset(
         dataset_name=dataset_name,
         fit_args=fit_args,
         expected_model_count=2,
