+++ tabular/tests/unittests/models/test_dummy.py
@@ -4,22 +4,23 @@ from pathlib import Path
 import pytest
 
 from autogluon.core.models.dummy.dummy_model import DummyModel
+from autogluon.tabular.testing import FitHelper, ModelFitHelper
 
 
-def test_no_models_will_raise(fit_helper, dataset_loader_helper):
+def test_no_models_will_raise():
     """Tests that RuntimeError is raised when no models fit"""
     fit_args = dict(
         hyperparameters={},
     )
 
     dataset_name = "toy_binary"
-    train_data, test_data, dataset_info = dataset_loader_helper.load_dataset(name=dataset_name)
+    train_data, test_data, dataset_info = FitHelper.load_dataset(name=dataset_name)
 
     with pytest.raises(RuntimeError):
-        fit_helper.fit_dataset(train_data=train_data, init_args=dict(label=dataset_info["label"]), fit_args=fit_args)
+        FitHelper.fit_dataset(train_data=train_data, init_args=dict(label=dataset_info["label"]), fit_args=fit_args)
 
 
-def test_no_models(fit_helper, dataset_loader_helper):
+def test_no_models():
     """Tests that logic works properly when no models are trained and raise_on_no_models_fitted=False"""
     fit_args = dict(
         hyperparameters={},
@@ -27,9 +28,9 @@ def test_no_models(fit_helper, dataset_loader_helper):
     )
 
     dataset_name = "toy_binary"
-    train_data, test_data, dataset_info = dataset_loader_helper.load_dataset(name=dataset_name)
+    train_data, test_data, dataset_info = FitHelper.load_dataset(name=dataset_name)
 
-    predictor = fit_helper.fit_dataset(train_data=train_data, init_args=dict(label=dataset_info["label"]), fit_args=fit_args)
+    predictor = FitHelper.fit_dataset(train_data=train_data, init_args=dict(label=dataset_info["label"]), fit_args=fit_args)
 
     assert not predictor.model_names()
     with pytest.raises(AssertionError):
@@ -39,7 +40,7 @@ def test_no_models(fit_helper, dataset_loader_helper):
     assert len(predictor.model_failures()) == 0
 
 
-def test_no_models_raise(fit_helper, dataset_loader_helper):
+def test_no_models_raise():
     """Tests that logic works properly when no models are trained, and tests predictor.model_failures() and raise_on_no_models_fitted=False"""
 
     expected_exc_str = "Test Error Message"
@@ -51,9 +52,9 @@ def test_no_models_raise(fit_helper, dataset_loader_helper):
     )
 
     dataset_name = "toy_binary"
-    train_data, test_data, dataset_info = dataset_loader_helper.load_dataset(name=dataset_name)
+    train_data, test_data, dataset_info = FitHelper.load_dataset(name=dataset_name)
 
-    predictor = fit_helper.fit_dataset(train_data=train_data, init_args=dict(label=dataset_info["label"]), fit_args=fit_args)
+    predictor = FitHelper.fit_dataset(train_data=train_data, init_args=dict(label=dataset_info["label"]), fit_args=fit_args)
 
     assert not predictor.model_names()
     with pytest.raises(AssertionError):
@@ -70,12 +71,12 @@ def test_no_models_raise(fit_helper, dataset_loader_helper):
     assert model_failures_dict["exc_str"] == expected_exc_str
 
 
-def test_raise_on_model_failure(fit_helper, dataset_loader_helper):
+def test_raise_on_model_failure():
     """Tests that logic works properly when model raises exception and raise_on_model_failure=True"""
 
     expected_exc_str = "Test Error Message"
 
-    train_data, test_data, dataset_info = dataset_loader_helper.load_dataset(name="toy_binary")
+    train_data, test_data, dataset_info = FitHelper.load_dataset(name="toy_binary")
 
     # Force DummyModel to raise an exception when fit.
     fit_args = dict(
@@ -85,37 +86,37 @@ def test_raise_on_model_failure(fit_helper, dataset_loader_helper):
     )
 
     with pytest.raises(ValueError) as excinfo:
-        fit_helper.fit_dataset(train_data=train_data, init_args=dict(label=dataset_info["label"]), fit_args=fit_args)
+        FitHelper.fit_dataset(train_data=train_data, init_args=dict(label=dataset_info["label"]), fit_args=fit_args)
     assert str(excinfo.value) == "Test Error Message"
 
 
-def test_dummy(fit_helper):
+def test_dummy():
     model_cls = DummyModel
     model_hyperparameters = {}
 
     """Additionally tests that all metrics work"""
-    fit_helper.verify_model(model_cls=model_cls, model_hyperparameters=model_hyperparameters, extra_metrics=True)
+    FitHelper.verify_model(model_cls=model_cls, model_hyperparameters=model_hyperparameters, extra_metrics=True)
 
 
-def test_dummy_binary_model(model_fit_helper):
+def test_dummy_binary_model():
     fit_args = dict()
     dataset_name = "toy_binary"
-    model_fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, model=DummyModel(), fit_args=fit_args)
+    ModelFitHelper.fit_and_validate_dataset(dataset_name=dataset_name, model=DummyModel(), fit_args=fit_args)
 
 
-def test_dummy_multiclass_model(model_fit_helper):
+def test_dummy_multiclass_model():
     fit_args = dict()
     dataset_name = "toy_multiclass"
-    model_fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, model=DummyModel(), fit_args=fit_args)
+    ModelFitHelper.fit_and_validate_dataset(dataset_name=dataset_name, model=DummyModel(), fit_args=fit_args)
 
 
-def test_dummy_regression_model(model_fit_helper):
+def test_dummy_regression_model():
     fit_args = dict()
     dataset_name = "toy_regression"
-    model_fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, model=DummyModel(), fit_args=fit_args)
+    ModelFitHelper.fit_and_validate_dataset(dataset_name=dataset_name, model=DummyModel(), fit_args=fit_args)
 
 
-def test_dummy_binary_absolute_path(fit_helper):
+def test_dummy_binary_absolute_path():
     """Test that absolute path works"""
     fit_args = dict(
         hyperparameters={DummyModel: {}},
@@ -126,10 +127,10 @@ def test_dummy_binary_absolute_path(fit_helper):
 
     dataset_name = "toy_binary"
 
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, init_args=init_args, fit_args=fit_args)
+    FitHelper.fit_and_validate_dataset(dataset_name=dataset_name, init_args=init_args, fit_args=fit_args)
 
 
-def test_dummy_binary_absolute_path_stack(fit_helper):
+def test_dummy_binary_absolute_path_stack():
     """Test that absolute path works"""
     fit_args = dict(
         hyperparameters={DummyModel: {}},
@@ -139,14 +140,14 @@ def test_dummy_binary_absolute_path_stack(fit_helper):
     )
 
     dataset_name = "toy_binary"
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args, expected_model_count=4, path_as_absolute=True)
+    FitHelper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args, expected_model_count=4, path_as_absolute=True)
 
 
-def test_dummy_binary_model_absolute_path(model_fit_helper):
+def test_dummy_binary_model_absolute_path():
     """Test that absolute path works"""
     fit_args = dict()
     path = Path(".") / "AG_test"
     path = str(path.resolve())
     model = DummyModel(path=path)
     dataset_name = "toy_binary"
-    model_fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, model=model, fit_args=fit_args)
+    ModelFitHelper.fit_and_validate_dataset(dataset_name=dataset_name, model=model, fit_args=fit_args)
