+++ tabular/tests/unittests/models/test_knn.py
@@ -1,8 +1,9 @@
 from autogluon.tabular.models.knn.knn_model import KNNModel
+from autogluon.tabular.testing import FitHelper
 
 
-def test_knn(fit_helper):
+def test_knn():
     model_cls = KNNModel
     model_hyperparameters = {"n_neighbors": 2}
 
-    fit_helper.verify_model(model_cls=model_cls, model_hyperparameters=model_hyperparameters)
+    FitHelper.verify_model(model_cls=model_cls, model_hyperparameters=model_hyperparameters)
