+++ tabular/tests/unittests/models/test_linear.py
@@ -1,8 +1,9 @@
 from autogluon.tabular.models.lr.lr_model import LinearModel
+from autogluon.tabular.testing import FitHelper
 
 
-def test_linear(fit_helper):
+def test_linear():
     model_cls = LinearModel
     model_hyperparameters = {}
 
-    fit_helper.verify_model(model_cls=model_cls, model_hyperparameters=model_hyperparameters)
+    FitHelper.verify_model(model_cls=model_cls, model_hyperparameters=model_hyperparameters)
