+++ tabular/tests/unittests/models/test_text_prediction_v1_model.py
@@ -1,9 +1,11 @@
 import pytest
 
+from autogluon.tabular.testing import FitHelper
+
 
 # FIXME: Post 0.1, refit_full by storing the iteration idx for reaching the best performance.
 @pytest.mark.gpu
-def test_text_prediction_v1_sts(fit_helper):
+def test_text_prediction_v1_sts():
     pytest.skip("Temporary skip the unittest")
     fit_args = dict(
         hyperparameters={
@@ -14,9 +16,9 @@ def test_text_prediction_v1_sts(fit_helper):
         },
     )
     dataset_name = "sts"
-    fit_helper.fit_and_validate_dataset(
+    FitHelper.fit_and_validate_dataset(
         dataset_name=dataset_name,
         fit_args=fit_args,
-        sample_size=1000,
+        sample_size=100,
         refit_full=True,
     )
