+++ tabular/tests/unittests/models/test_vowpalwabbit.py
@@ -3,30 +3,31 @@ import sys
 import pytest
 
 from autogluon.tabular.models.vowpalwabbit.vowpalwabbit_model import VowpalWabbitModel
+from autogluon.tabular.testing import FitHelper
 
 
 @pytest.mark.skipif(sys.version_info >= (3, 11) or sys.platform=="darwin", reason="vowpalwabbit doesn't support python 3.11 and above or MacOS yet.")
-def test_vowpalwabbit_binary(fit_helper):
+def test_vowpalwabbit_binary():
     fit_args = dict(
         hyperparameters={VowpalWabbitModel: {}},
     )
     dataset_name = "adult"
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
+    FitHelper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
 
 
 @pytest.mark.skipif(sys.version_info >= (3, 11) or sys.platform=="darwin", reason="vowpalwabbit doesn't support python 3.11 and above or MacOS yet.")
-def test_vowpalwabbit_multiclass(fit_helper):
+def test_vowpalwabbit_multiclass():
     fit_args = dict(
         hyperparameters={VowpalWabbitModel: {}},
     )
     dataset_name = "covertype_small"
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
+    FitHelper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
 
 
 @pytest.mark.skipif(sys.version_info >= (3, 11) or sys.platform=="darwin", reason="vowpalwabbit doesn't support python 3.11 and above or MacOS yet.")
-def test_vowpalwabbit_regression(fit_helper):
+def test_vowpalwabbit_regression():
     fit_args = dict(
         hyperparameters={VowpalWabbitModel: {}},
     )
     dataset_name = "ames"
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
+    FitHelper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
